+++ files/settings/Groovy.groovy
@@ -1,32 +1,24 @@
 #!groovy
 import hudson.plugins.groovy.*
-import jenkins.model.*
-import hudson.tools.*
+import hudson.tools.InstallSourceProperty
 import java.util.logging.Level
 import java.util.logging.Logger
 
 Logger logger = Logger.getLogger('jenkins.instance.settings.groovy')
-
-def desc = Jenkins.getInstance().getDescriptor("hudson.plugins.groovy.GroovyInstallation")
-
-def prev = Jenkins.getInstance().getDescriptorByType(Groovy.DescriptorImpl.class).getInstallations()
-def exists = false
+def groovyPlugin = Jenkins.getInstance().getExtensionList(hudson.plugins.groovy.Groovy.DescriptorImpl.class)[0]
 
 def groovyName = "Groovy_${version}"
-def installations = []
 
-prev.each{ groovy ->
-  if (groovy.getName() == groovyName) {
-    logger.log(Level.INFO, groovy.getName() + " already exists")
-    exists = true
-  }
-  installations.push(groovy)
+def groovyInstall = groovyPlugin.installations.find {
+  install -> install.name.equals(groovyName)
 }
-if (exists) {
+
+if (groovyInstall != null) {
+  logger.log(Level.INFO, groovyName + " already exists")
   return
 }
 
 def installer = new GroovyInstaller("${version}")
 def prop = new InstallSourceProperty([installer])
-installations.add(new GroovyInstallation(groovyName, "", [prop]))
-desc.setInstallations(*installations)
+groovyPlugin.installations += new GroovyInstallation(groovyName, null, [prop])
+groovyPlugin.save()