+++ files/settings/Java.groovy
@@ -8,33 +8,25 @@ import java.util.logging.Logger
 
 Logger logger = Logger.getLogger('jenkins.instance.settings.java')
 
-def jdkDescriptor = Jenkins.getInstance().getDescriptor("hudson.model.JDK")
+def javaPlugin = Jenkins.getInstance().getExtensionList(hudson.model.JDK.DescriptorImpl.class)[0]
 def jdkInstallerDescriptor = Jenkins.getInstance().getDescriptor("hudson.tools.JDKInstaller")
-
-def installations = Jenkins.getInstance().getJDKs()
-def exists = false
-
+def jdks = Jenkins.getInstance().getJDKs()
 def jdkName = "java${version}-${implementation}"
 
-installations.find{ jdk ->
-  if (jdk.getName() == jdkName) {
-    logger.log(Level.INFO, jdk.getName() + " already exists")
-    exists = true
-    return true
-  }
-  return false
+def javaInstall = javaPlugin.installations.find {
+  install -> install.name.equals(jdkName)
 }
-if (exists) {
+
+if (javaInstall != null) {
+  logger.log(Level.INFO, jdkName + " already exists")
   return
 }
 
 if ("${implementation}" == "openjdk") {
-  installations.push(new JDK(jdkName, "${home}"))
+  jdks.push(new JDK(jdkName, "${home}"))
 } else if ("${implementation}" == "oraclejdk") {
   def installer = new JDKInstaller("${install}", true)
   def installerProps = new InstallSourceProperty([installer])
-  installations.push(new JDK(jdkName, null, [installerProps]))
+  jdks.push(new JDK(jdkName, null, [installerProps]))
   jdkInstallerDescriptor.doPostCredential("${user}", "${password}")
 }
-
-jdkDescriptor.setInstallations(installations.toArray(new JDK[0]))