+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a changelog](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased](https://github.com/idealista/jenkins-role/tree/develop)
+### Changed
+- *Install default plugins* @jmonterrubio
+- *Install default settings* @jmonterrubio
+- *Remove CLI* @jmonterrubio
+
+### Added
+- *Improve security* @jmonterrubio
+- *Job DSL execution* @jmonterrubio
+
 
 ## [1.2.0](https://github.com/idealista/jenkins-role/tree/1.2.0)
 ## [Full Changelog](https://github.com/idealista/jenkins-role/compare/1.1.0...1.2.0)